Add files via upload
This will create and update a map marker where the gunner & commander are looking for every crew member of a given vehicle.
It will also create a cone of vision, based on their current FOV.

HOW TO USE:
Add this script to your mission's folder, and in the init of your object add the following
myvehicle execVM "CrewVisionMap.sqf";

"myvehicle" must be the variable name of the vehicle you wish to initialise this script for. "this" cannot be used as an argument.
To initialise for more than one vehicle, simply copy the init code and change "myvehicle" to the name of your new vehicle.

DO NOT EXECUTE SERVER-ONLY, or it will not work.